+++ db/migrate/2013040400035_create_stage_populations.rb
@@ -3,7 +3,7 @@ class CreateStagePopulations < ActiveRecord::Migration
     create_table :stage_populations do |t|
       t.integer :citation_id
       t.integer :stage_id
-      t.float :population
+      t.decimal :population
       t.text :comment
       t.string :datum
       t.integer :user_id
