+++ db/migrate/2013040400036_create_stage_prod_biomass_ratios.rb
@@ -3,7 +3,7 @@ class CreateStageProdBiomassRatios < ActiveRecord::Migration
     create_table :stage_prod_biomass_ratios do |t|
       t.integer :citation_id
       t.integer :stage_id
-      t.float :prod_biomass_ratio
+      t.decimal :prod_biomass_ratio
       t.text :comment
       t.string :datum
       t.integer :user_id
